// src/app/models/selcom.model.ts
export class Selcom {
    selcom_id: number;
    transid: string;
    order_id: string;
    reference: string;
    result: string;
    resultcode: string;
    payment_status: string;
    user_id: number;
    specialist_id: number;
    amount: string;
    phonenumber_used: string;
    type: string;
    payment_type: string;
    client: {
        userID: number;
        firstName: string;
        lastName: string;
        dateOfBirth: string;
        email: string;
        password: string;
        phone: string;
        gender: string;
        age: number
        regionID: string;
        street: string;
        lat: string;
        longt: string;
        geo: string;
        timeSt: string;
        doctorsIDnumber: string;
        doctorsIDverificationStatus: string;
        status: string;
        userRole: number
        userNhifNumber: string;
        onlineStatus: number;
        specializationID: number;
        lastOnlineTime: string;
        user_image: string;
        height: string;
        weight: string;
        blood_group: string;
        doctorPromotionCode: string;
        userPromotionCode: string;
        allergy: string;
        doctor_bio: string;
        specilizationAreaID: number;
        experience: string;
        consultation_fee: number;
        no_initial_consultation: string;
        no_remain_consulatation: string;
        consultation_availabiliy: string;
        phone_availability: string;
        patient_no: string;
        ip_address: string;
        deleted: string;
    };
    doctor: {
        userID: number;
        firstName: string;
        lastName: string;
        dateOfBirth: string;
        email: string;
        password: string;
        phone: string;
        gender: string;
        age: number
        regionID: string;
        street: string;
        lat: string;
        longt: string;
        geo: string;
        timeSt: string;
        doctorsIDnumber: string;
        doctorsIDverificationStatus: string;
        status: string;
        userRole: number
        userNhifNumber: string;
        onlineStatus: number;
        specializationID: number;
        lastOnlineTime: string;
        user_image: string;
        height: string;
        weight: string;
        blood_group: string;
        doctorPromotionCode: string;
        userPromotionCode: string;
        allergy: string;
        doctor_bio: string;
        specilizationAreaID: number;
        experience: string;
        consultation_fee: number;
        no_initial_consultation: string;
        no_remain_consulatation: string;
        consultation_availabiliy: string;
        phone_availability: string;
        patient_no: string;
        ip_address: string;
        deleted: string;
    }
}